More generic components
The structure material_list_t is replaced by components_t. It is a list
of materials and items, which can be given in different amounts and can
be kept or consumed during the work process. For now, components_t is also
used to represent the products (here, .consumed is unused). Moreover, the
current implementation of components_t only works with materials; handling
items will require an inventory (storage of both materials and items).