Fixing pixel decimal coordinates, again
The half-coordinate hack of translating by (0.375,0.375) is irrelevant in
SFML views and should be handled by the library. However, when mapping
texture in the map's vertex array, glitch still occurs, probably due
to rounding errors. This is is fixed by cropping a tiny fraction of the
external pixels of the texture.