Unifying skills and bonuses
There used to be special, material and item skills and bonuses. There
are now only 'skills' or 'bonuses'. Special skills are the first ones,
so that their id is unchanged. Material and item skills are indicated
in the kindOf_{material,item}_t structure. The character can now handle
all skills and bonuses in one method.