Replacing main() by WinMain() on Windows
This commit allows to compile the program as a Windows GUI application
using the linker flag '-mwindows'. It is done so that the application
started from the graphical interface will not open a console but, if
started from a console, it will still be able to write to it. It is
actually the default behaviour on Linux-based systems.